# docs/swagger.yaml
basePath: /v1/api/
definitions:
  common.ValidationResponse:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  controller.questionVoteInput:
    properties:
      negative:
        type: boolean
    type: object
  entity.GameStatus:
    enum:
    - created
    - started
    - ended
    type: string
    x-enum-varnames:
    - GameStatusCreated
    - GameStatusStarted
    - GameStatusEnded
  model.Answer:
    properties:
      answer_id:
        type: string
      is_correct:
        type: boolean
      label:
        type: string
      question_id:
        type: string
    required:
    - answer_id
    - is_correct
    - label
    - question_id
    type: object
  model.Auth:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.CreateQuestionInput:
    properties:
      correct_answers:
        items:
          type: string
        type: array
      label:
        type: string
      wrong_answers:
        items:
          type: string
        type: array
    required:
    - correct_answers
    - label
    - wrong_answers
    type: object
  model.Game:
    properties:
      created_at:
        type: string
      end_at:
        type: string
      game_id:
        type: string
      players:
        items:
          $ref: '#/definitions/model.User'
        type: array
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      started_at:
        type: string
      status:
        $ref: '#/definitions/entity.GameStatus'
    required:
    - created_at
    - game_id
    - players
    - questions
    - status
    type: object
  model.GameResult:
    properties:
      created_at:
        type: string
      end_at:
        type: string
      game_id:
        type: string
      players:
        items:
          $ref: '#/definitions/model.User'
        type: array
      question_results:
        items:
          $ref: '#/definitions/model.QuestionResult'
        type: array
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
      started_at:
        type: string
      status:
        $ref: '#/definitions/entity.GameStatus'
    required:
    - created_at
    - game_id
    - players
    - question_results
    - questions
    - status
    type: object
  model.GameStatus:
    properties:
      exists:
        type: boolean
      is_user_in_game:
        type: boolean
      status:
        $ref: '#/definitions/entity.GameStatus'
    required:
    - exists
    - is_user_in_game
    type: object
  model.GeneralResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    required:
    - code
    - data
    - message
    type: object
  model.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/model.Answer'
        type: array
      created_by:
        $ref: '#/definitions/model.User'
      difficulty:
        type: integer
      label:
        type: string
      question_id:
        type: string
      vote_count:
        type: integer
      vote_sum:
        type: integer
    required:
    - answers
    - difficulty
    - label
    - question_id
    - vote_count
    - vote_sum
    type: object
  model.QuestionResult:
    properties:
      answers:
        items:
          $ref: '#/definitions/model.Answer'
        type: array
      created_by:
        $ref: '#/definitions/model.User'
      difficulty:
        type: integer
      label:
        type: string
      question_id:
        type: string
      user_answers:
        items:
          $ref: '#/definitions/model.UserAnswer'
        type: array
      vote_count:
        type: integer
      vote_sum:
        type: integer
    required:
    - answers
    - difficulty
    - label
    - question_id
    - vote_count
    - vote_sum
    type: object
  model.User:
    properties:
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
    required:
    - user_id
    - username
    type: object
  model.UserAnswer:
    properties:
      answer:
        $ref: '#/definitions/model.Answer'
      game_id:
        type: string
      is_correct:
        type: boolean
      question_id:
        type: string
      text:
        type: string
      user_answer_id:
        type: string
      user_id:
        type: string
    required:
    - question_id
    - user_answer_id
    - user_id
    type: object
  model.UserLoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.UserSignupInput:
    properties:
      email:
        type: string
      password:
        maxLength: 30
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 4
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8000
info:
  contact: {}
  description: REST API for creating quiz
  title: GoQuizz Swagger API
  version: "1.0"
paths:
  /game:
    get:
      consumes:
      - application/json
      description: |-
        Get all games for current user
        Access restricted to: USER (as a player) or ADMIN
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Game'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get all games for current user
      tags:
      - game
    post:
      consumes:
      - application/json
      description: |-
        User can create a game, wait for his friends to join (JoinGame), and start the game.
        Access restricted to: USER
      parameters:
      - description: 'todo: game parameters'
        in: body
        name: input
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Game'
              type: object
      security:
      - Bearer: []
      summary: Create a new game
      tags:
      - game
  /game/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Get a game by ID
        Access restricted to: USER
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Game'
              type: object
      security:
      - Bearer: []
      summary: Get a game by ID
      tags:
      - game
  /game/{id}/answer:
    post:
      consumes:
      - application/json
      description: |-
        User can answer a question in a game if the game is started and the question is not already answered.
        Access restricted to: USER
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      - description: question Id
        in: path
        name: id
        required: true
        type: string
      - description: 'todo: answer info'
        in: body
        name: input
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - Bearer: []
      summary: Answer a question in a game
      tags:
      - game
  /game/{id}/join:
    patch:
      consumes:
      - application/json
      description: |-
        User can join an existing game if not already a player of it, and if the game is not started.
        Access restricted to: USER
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Game'
              type: object
      security:
      - Bearer: []
      summary: Join an existing game
      tags:
      - game
  /game/{id}/result:
    get:
      consumes:
      - application/json
      description: |-
        Get a game result by game ID
        Access restricted to: USER
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GameResult'
              type: object
      summary: Get a game result by game ID
      tags:
      - game
  /game/{id}/start:
    patch:
      consumes:
      - application/json
      description: |-
        One of the players can start an existing game if game is not already started.
        Access restricted to: USER
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Game'
              type: object
      security:
      - Bearer: []
      summary: Start an existing game
      tags:
      - game
  /game/{id}/status:
    get:
      consumes:
      - application/json
      description: |-
        Get a game status by game ID
        Access not restricted
      parameters:
      - description: game Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.GameStatus'
              type: object
      summary: Get a game status by game ID
      tags:
      - game
  /login:
    post:
      consumes:
      - application/json
      description: Verify provided credentials and return JWT token + user associated
      parameters:
      - description: sign in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Auth'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/common.ValidationResponse'
                  type: array
              type: object
      summary: User Login
      tags:
      - user
  /question:
    post:
      consumes:
      - application/json
      description: |-
        Create a new question, associated to the current user
        Access restricted to: ADMIN
      parameters:
      - description: question and answers info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateQuestionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Question'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/common.ValidationResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Create a question
      tags:
      - question
  /question/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Get a question by ID
        Access restricted to: USER
      parameters:
      - description: question id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Question'
              type: object
      security:
      - Bearer: []
      summary: Get a question by ID
      tags:
      - question
  /question/{id}/vote:
    put:
      consumes:
      - application/json
      description: |-
        User can upvote/downvote a question. Downvote are done by passing the "negative" prop in the request body.
        It will upsert the vote -> Creating if user/question relation does not exist, update it otherwise.
        Access restricted to: USER
      parameters:
      - description: question id
        in: path
        name: id
        required: true
        type: string
      - description: question vote info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.questionVoteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: User vote for a question
      tags:
      - question
  /refresh:
    post:
      consumes:
      - application/json
      description: Verify the Authorization token, refresh it if valid, return null
        user otherwise
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Auth'
              type: object
      summary: User Refresh
      tags:
      - user
  /signup:
    post:
      consumes:
      - application/json
      description: Create user and return JWT token + user associated
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserSignupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Auth'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/common.ValidationResponse'
                  type: array
              type: object
      summary: User Sign up
      tags:
      - user
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    scopes:
      admin: ' Grants full read and write access'
      mod: ' Grants partial read and write access'
      user: ' Only access/update his data'
    type: apiKey
swagger: "2.0"